requirements.txt
numpy==1.19.5
pytest-cov==2.8.1
sklearn==0.0
pytest==6.2.2
#allure-pytest==2.7.0
#pytest-allure-adaptor==1.7.10
pytest-assume==2.4.3
pytest-timeout==1.3.3
pytest-repeat==0.8.0
allure-pytest==2.7.0
pytest-print==0.2.1
pytest-level==0.1.1
pytest-xdist==2.2.1
pytest-rerunfailures==9.1.1
git+https://github.com/Projectplace/pytest-tags
ndg-httpsclient
pyopenssl
pyasn1
pytest-html==3.1.1
delayed-assert
#pytest-sugar==0.9.4
pytest-reportportal==5.0.10
pytest-parallel
pytest-random-order

# for customize config test
python-benedict==0.24.3

# version need to be consistent with protobuf used in pymilvus
protobuf==3.17.1

#grpcio==1.37.1
#grpcio-testing==1.37.1
#grpcio-tools==1.37.1
pandas==1.2.4  # 1.1.5
h5py==3.3.0  # 2.7.1
influxdb-client==1.21.0
PyYAML==5.4.1
tableprint==0.8.0
ansicolors==1.1.8
kubernetes==17.17.0
locust>=1.3.2
pymongo==3.10.0
apscheduler==3.7.0
pyhelm==2.14.5
timeout_decorator==0.5.0

requests==2.26.0
urllib3==1.25.11
PyMySQL==1.0.2
